Pulled the hand-built topological molecule creation functions from test_forcefield.py into create_molecules.py so other tests can use them without reaching into test_forcefield.py. This modularization also reduces import time overhead during manual testing.
